+++ src/Text/Show/Text/Data/Data.hs
@@ -28,8 +28,8 @@ import Prelude hiding (Show)
 import Text.Show.Text.Classes (Show(showb, showbPrec))
 import Text.Show.Text.Data.List ()
 import Text.Show.Text.Data.Ratio ()
-import Text.Show.Text.TH.Internal (deriveShowPragmas, defaultInlineShowbPrec,
-                                   defaultInlineShowb)
+import Text.Show.Text.TH.Internal (deriveShow, deriveShowPragmas,
+                                   defaultInlineShowbPrec, defaultInlineShowb)
 
 #include "inline.h"
 
@@ -69,8 +69,8 @@ showbConstrRepPrec = showbPrec
 {-# INLINE showbConstrRepPrec #-}
 
 $(deriveShowPragmas defaultInlineShowbPrec ''DataType)
-$(deriveShowPragmas defaultInlineShowbPrec ''DataRep)
-$(deriveShowPragmas defaultInlineShowbPrec ''ConstrRep)
+$(deriveShow                               ''DataRep)
+$(deriveShow                               ''ConstrRep)
 $(deriveShowPragmas defaultInlineShowb     ''Fixity)
 
 instance Show Constr where