+++ src/Text/Show/Text/GHC/Fingerprint.hs
@@ -14,6 +14,7 @@ This module is only available with @base-4.4.0.0@ or later.
 -}
 module Text.Show.Text.GHC.Fingerprint (showbFingerprint) where
 
+import Data.Semigroup (timesN)
 import Data.Text.Lazy.Builder (Builder)
 import Data.Word (Word64)
 
@@ -23,7 +24,7 @@ import Prelude hiding (Show)
 
 import Text.Show.Text.Classes (Show(showb))
 import Text.Show.Text.Data.Integral (showbHex)
-import Text.Show.Text.Utils ((<>), lengthB, replicateB, s)
+import Text.Show.Text.Utils ((<>), lengthB, s)
 
 -- | Convert a 'Fingerprint' to a 'Builder'.
 -- This function is only available with @base-4.4.0.0@ or later.
@@ -34,8 +35,7 @@ showbFingerprint (Fingerprint w1 w2) = hex16 w1 <> hex16 w2
   where
     hex16 :: Word64 -> Builder
     hex16 i = let hex = showbHex i
-               in replicateB (16 - lengthB hex) (s '0') <> hex
-{-# INLINE showbFingerprint #-}
+              in timesN (fromIntegral . max 0 $ 16 - lengthB hex) (s '0') <> hex
 
 instance Show Fingerprint where
     showb = showbFingerprint