+++ tests/Properties/Builder.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE CPP #-}
 {-|
 Module:      Properties.Builder
 Copyright:   (C) 2014-2015 Ryan Scott
@@ -18,19 +17,10 @@ import Test.Tasty.QuickCheck (testProperty)
 
 import Text.Show.Text (Builder, fromString, fromText, lengthB,
                        toString, toText, unlinesB, unwordsB)
-#if !defined(mingw32_HOST_OS)
-import Text.Show.Text (replicateB, singleton)
-#endif
 
 -- | Verifies 'lengthB' and 'length' produce the same output.
 prop_lengthB :: String -> Bool
 prop_lengthB s = fromIntegral (lengthB $ fromString s) == length s
-
-#if !defined(mingw32_HOST_OS)
--- | Verifies 'replicateB' and 'replicate' produce the same output.
-prop_replicateB :: Int -> Char -> Bool
-prop_replicateB i c = replicateB (fromIntegral i) (singleton c) == fromString (replicate i c)
-#endif
 
 -- | Verifies @fromText . toText = id@.
 prop_toText :: Builder -> Bool
@@ -52,10 +42,6 @@ builderTests :: [TestTree]
 builderTests =
     [ testGroup "Builder-related functions"
         [ testProperty "lengthB output"             prop_lengthB
-#if !defined(mingw32_HOST_OS)
--- TODO: Figure out why this diverges on Windows
-        , testProperty "replicateB output"          prop_replicateB
-#endif
         , testProperty "fromString . toString = id" prop_toString
         , testProperty "fromText . toText = id"     prop_toText
         , testProperty "unlinesB output"            prop_unlinesB